task-5-websockets/requirements.txt
websockets
asyncio
sqlalchemy
psycopg2-binary
python-dotenv